+++ test/endpoint/search.js
@@ -56,6 +56,28 @@ describe('search', function () {
       should.not.exist(checks.handler_err);
       checks.handler_result.foo.should.equal('bar');
     });
+
+    it('should pass result to callback when search modified handler is called', function (done) {
+      function cb(err, result) {
+        checks.handler_err = err;
+        checks.handler_result = result;
+        done();
+      }
+      search.handlers(cb)['206']({ foo: 'bar' });
+      should.not.exist(checks.handler_err);
+      checks.handler_result.foo.should.equal('bar');
+    });
+
+    it('should pass error and result to callback when validation error handler is called', function (done) {
+      function cb(err, result) {
+        checks.handler_err = err;
+        checks.handler_result = result;
+        done();
+      }
+      search.handlers(cb)['400']({ foo: 'bar' });
+      checks.handler_err.message.should.equal('Validation error');
+      checks.handler_result.foo.should.equal('bar');
+    });
   });
 });